[Windows, tests] Fix most of the test failures on Windows Server 2019 VM. There are few flaky tests left

+++ pax-web-itest/pax-web-itest-container/pax-web-itest-container-common/src/main/java/org/ops4j/pax/web/itest/container/whiteboard/AbstractSimultaneousWhiteboardIntegrationTest.java
@@ -45,7 +45,9 @@ public abstract class AbstractSimultaneousWhiteboardIntegrationTest extends Abst
 	protected Option[] baseConfigure() {
 		File dir = new File("target/bundles");
 		try {
-			FileUtils.cleanDirectory(dir);
+			if (dir.isDirectory()) {
+				FileUtils.cleanDirectory(dir);
+			}
 			dir.mkdirs();
 			InputStream b1 = TinyBundles.bundle().add(TestActivator.class)
 					.add(WhiteboardFilter.class)
